+++ packages/desktop/src/renderer/components/Channel/CreateChannel/CreateChannel.tsx
@@ -19,7 +19,7 @@ export const CreateChannel = () => {
   const channels = useSelector(publicChannels.selectors.publicChannels)
 
   const communityErrors = useSelector(errors.selectors.currentCommunityErrors)
-  const error = communityErrors[SocketActionTypes.CREATED_CHANNEL]
+  const error = communityErrors[SocketActionTypes.CREATE_CHANNEL]
 
   const createChannelModal = useModal(ModalName.createChannel)
 
@@ -49,7 +49,7 @@ export const CreateChannel = () => {
       console.error('No identity found')
       dispatch(
         errors.actions.addError({
-          type: SocketActionTypes.CREATED_CHANNEL,
+          type: SocketActionTypes.CREATE_CHANNEL,
           code: ErrorCodes.NOT_FOUND,
           message: ErrorMessages.GENERAL,
           community: community,
@@ -61,7 +61,7 @@ export const CreateChannel = () => {
     if (channels.some(channel => channel.name === name)) {
       dispatch(
         errors.actions.addError({
-          type: SocketActionTypes.CREATED_CHANNEL,
+          type: SocketActionTypes.CREATE_CHANNEL,
           code: ErrorCodes.FORBIDDEN,
           message: ErrorMessages.CHANNEL_NAME_TAKEN,
           community: community,